StreamlitAPP.py
import os
import json
import traceback
import pandas as pd
from dotenv import load_dotenv,find_dotenv
from src.mcqgenerator.utils import read_file,get_table_data
import streamlit as st
from src.mcqgenerator.MCQGenerator import generate_evaluate_chain
from src.mcqgenerator.logger import logging

