requirements.txt
openai
langchain==0.3.13
streamlit
python-dotenv
PyPDF2
langchain-ai21

-e .